Stub test generator behavior within tests
Remove some individual implementations of `Generator` and
`GeneratorForAnnotation` in tests in favor of a general `_StubGenerator`
implementation which can take a callback for it's behavior. This keeps
the interesting behavior for the test contained within the test.

Move the test for `GeneratorForAnnotation` not resolving libraries which
have no annotated elements into the relevant test suite.
