+++ source_gen/CHANGELOG.md
@@ -1,3 +1,5 @@
+## 1.1.2-dev
+
 ## 1.1.1
 
 - Drop dependency on `package:pedantic`.
