+++ source_gen/pubspec.yaml
@@ -1,5 +1,5 @@
 name: source_gen
-version: 1.1.1
+version: 1.1.2-dev
 description: >-
   Source code generation builders and utilities for the Dart build system
 repository: https://github.com/dart-lang/source_gen
