+++ source_gen/test/generator_for_annotation_test.dart
@@ -4,6 +4,8 @@
 
 // The first test that runs `testBuilder` takes a LOT longer than the rest.
 @Timeout.factor(3)
+import 'package:analyzer/dart/analysis/utilities.dart';
+import 'package:analyzer/dart/ast/ast.dart';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:build/build.dart';
 import 'package:build_test/build_test.dart';
@@ -20,7 +22,8 @@ void main() {
       'list with null, empty, and whitespace items': [null, '', '\n \t']
     }.entries) {
       test(entry.key, () async {
-        final generator = LiteralOutput(entry.value);
+        final generator =
+            _StubGenerator<Deprecated>('Value', (_) => entry.value);
         final builder = LibraryBuilder(generator);
         await testBuilder(builder, _inputMap, outputs: {});
       });
@@ -28,14 +31,17 @@ void main() {
   });
 
   test('Supports and dedupes multiple return values', () async {
-    const generator = RepeatingGenerator();
+    final generator = _StubGenerator<Deprecated>('Repeating', (element) sync* {
+      yield '// There are deprecated values in this library!';
+      yield '// ${element.name}';
+    });
     final builder = LibraryBuilder(generator);
     await testBuilder(builder, _inputMap, outputs: {
       'a|lib/file.g.dart': r'''
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
 // **************************************************************************
-// RepeatingGenerator
+// Generator: Repeating
 // **************************************************************************
 
 // There are deprecated values in this library!
@@ -51,8 +57,13 @@ void main() {
 
   group('handles errors correctly', () {
     for (var entry in {
-      'sync errors': const FailingGenerator(),
-      'from iterable': const FailingIterableGenerator()
+      'sync errors': _StubGenerator<Deprecated>('Failing', (_) {
+        throw StateError('not supported!');
+      }),
+      'from iterable': _StubGenerator<Deprecated>('FailingIterable', (_) sync* {
+        yield '// There are deprecated values in this library!';
+        throw StateError('not supported!');
+      })
     }.entries) {
       test(entry.key, () async {
         final builder = LibraryBuilder(entry.value);
@@ -70,53 +81,38 @@ void main() {
       });
     }
   });
+
+  test('Does not resolve the library if there are no top level annotations',
+      () async {
+    final builder =
+        LibraryBuilder(_StubGenerator<Deprecated>('Deprecated', (_) {}));
+    final input = AssetId('a', 'lib/a.dart');
+    final assets = {input: 'main() {}'};
+
+    final reader = InMemoryAssetReader(sourceAssets: assets);
+    final resolver = _TestingResolver(assets);
+
+    await runBuilder(builder, [input], reader, InMemoryAssetWriter(),
+        _FixedResolvers(resolver));
+
+    expect(resolver.parsedUnits, {input});
+    expect(resolver.resolvedLibs, isEmpty);
+  });
 }
 
-class FailingIterableGenerator extends GeneratorForAnnotation<Deprecated> {
-  const FailingIterableGenerator();
+class _StubGenerator<T> extends GeneratorForAnnotation<T> {
+  final String _name;
+  final Object? Function(Element) _behavior;
+
+  const _StubGenerator(this._name, this._behavior);
 
   @override
-  Iterable<String> generateForAnnotatedElement(
-      Element element, ConstantReader annotation, BuildStep buildStep) sync* {
-    yield '// There are deprecated values in this library!';
-    throw StateError('not supported!');
-  }
-
-  @override
-  String toString() => 'FailingGenerator';
-}
-
-class FailingGenerator extends GeneratorForAnnotation<Deprecated> {
-  const FailingGenerator();
-
-  @override
-  dynamic generateForAnnotatedElement(
-      Element element, ConstantReader annotation, BuildStep buildStep) {
-    throw StateError('not supported!');
-  }
-}
-
-class RepeatingGenerator extends GeneratorForAnnotation<Deprecated> {
-  const RepeatingGenerator();
-
-  @override
-  Iterable<String> generateForAnnotatedElement(
-      Element element, ConstantReader annotation, BuildStep buildStep) sync* {
-    yield '// There are deprecated values in this library!';
-
-    yield '// ${element.name}';
-  }
-}
-
-class LiteralOutput<T> extends GeneratorForAnnotation<Deprecated> {
-  final T? value;
-
-  const LiteralOutput([this.value]);
-
-  @override
-  T? generateForAnnotatedElement(
+  Object? generateForAnnotatedElement(
           Element element, ConstantReader annotation, BuildStep buildStep) =>
-      null;
+      _behavior(element);
+
+  @override
+  String toString() => _name;
 }
 
 const _inputMap = {
@@ -131,3 +127,50 @@ const _inputMap = {
      final baz = 'baz';
      '''
 };
+
+class _TestingResolver implements ReleasableResolver {
+  final Map<AssetId, String> assets;
+  final parsedUnits = <AssetId>{};
+  final resolvedLibs = <AssetId>{};
+
+  _TestingResolver(this.assets);
+
+  @override
+  Future<CompilationUnit> compilationUnitFor(AssetId assetId,
+      {bool allowSyntaxErrors = false}) async {
+    parsedUnits.add(assetId);
+    return parseString(content: assets[assetId]!).unit;
+  }
+
+  @override
+  Future<bool> isLibrary(AssetId assetId) async {
+    final unit = await compilationUnitFor(assetId);
+    return unit.directives.every((d) => d is! PartOfDirective);
+  }
+
+  @override
+  Future<LibraryElement> libraryFor(AssetId assetId,
+      {bool allowSyntaxErrors = false}) async {
+    resolvedLibs.add(assetId);
+    return null as LibraryElement;
+  }
+
+  @override
+  void release() {}
+
+  @override
+  void noSuchMethod(_) => throw UnimplementedError();
+}
+
+class _FixedResolvers implements Resolvers {
+  final ReleasableResolver _resolver;
+
+  _FixedResolvers(this._resolver);
+
+  @override
+  Future<ReleasableResolver> get(BuildStep buildStep) =>
+      Future.value(_resolver);
+
+  @override
+  void reset() {}
+}
